Time: 7 ms (41.96%), Space: 13.8 MB (87.72%) - LeetHub

+++ 2063-vowels-of-all-substrings/2063-vowels-of-all-substrings.cpp
@@ -5,8 +5,8 @@ public:
         long long cnt=0;
         for(int i=0;i<n;i++){
             if(word[i]=='a' || word[i]=='e' || word[i]=='i' || word[i]=='o' or word[i]=='u'){
-                long long left=i+1;
-                long long right=(n-i);
+                long long left=(long long)i+1;
+                long long right=(long long)(n-i);
                 cnt+=(left*right);
             }
         }
